Test: switch size-overflow repro to Int64
Int is 4 bytes on wasm32, so 1_000_000_000 elements fits under UINT32_MAX
there and the diagnostic correctly does not fire. Use Int64 (always 8
bytes) so the test overflows on every pointer-size target.

+++ test/IRGen/inline_array_size_overflow.swift
@@ -5,8 +5,12 @@
 // whose total size exceeds 2^32 bytes cannot be represented. Reject these
 // eagerly instead of taking time proportional to the array's byte size in
 // IRGen.
+//
+// Use `Int64` rather than `Int` for the element type so the total overflows
+// on 32-bit targets as well (where `Int` is 4 bytes); 1_000_000_000 × 8 =
+// ~8 GiB either way.
 
 func tooLarge() {
-  _ = InlineArray<1_000_000_000, Int>(repeating: 0)
+  _ = InlineArray<1_000_000_000, Int64>(repeating: 0)
 }
-// CHECK: error: size of value of type 'Builtin.FixedArray<1000000000, Int>' exceeds the maximum supported size of 4 GiB
+// CHECK: error: size of value of type 'Builtin.FixedArray<1000000000, Int64>' exceeds the maximum supported size of 4 GiB
